Introduce configurable worker directory in Dockerfile
This commit allows the worker directory to be defined through an ARG directive in the Dockerfile. It modifies the build and copy steps to use the new WORKER_DIR argument value. Additionally, a "migrate" script has been added to the package.json file. This improves the flexibility and configurability of the build process.