Refactor schema component and add new data schemas
The schema component of the application has been refactored. The old index file was deleted as new schema files for `address`, `assessment`, `orders`, and `products` have been introduced. The new schema structure helps to manage data more effectively and easily. The build and typescript configuration for this component was also updated. Changes also include updated scripts and dependencies in the component's package.json.